# perf: improve gRPC query performance