Potentially might fix (but not in gh sense): https://github.com/influxdata/conductor/issues/264

This PR improves the performance of `read_filter`, `read_group` and `window_aggregate` gRPC operations by reducing the use of buffered channels in the exeution stage of the query engine.

### Benchmark

I wrote two benchmarks in this PR benchmarking the performance of `read_filter` and `read_group`. Both of them exercise the `influxrpc` frontend by effectively writing some data into in-memory chunks (across different configurations in the MUB/RB) and then planning and executing a query against those data.

The test data is 10,000 line protocol points comprising five tag keys with cardinalities of `{2, 10, 10, 50 1}`. The total cardinality is 10,000.

I then effectively exercised the equivalent of the following InfluxQL:

```
// read_filter queries
SELECT * FROM m0;                          // "no_pred"
SELECT * FROM m0 WHERE "tag3" = 'value49'; // "with_pred"

// read_group queries
SELECT SUM(v0) FROM m0 GROUP BY "tag2";                          // "no_pred"
SELECT SUM(v0) FROM m0 WHERE "tag3" = 'value49' GROUP BY "tag2"; // "with_pred"
``` 

Each of the queries above will return either `~10,000` or `~200` individual data frames, which would be sent as a response on the gRPC API. The current implementation in IOx has all these frames effectively flowing through some channels with small fixed buffers, which causes contention over these channels and also lots of context switching across threads/tasks.

Here is an example of a profile of one of the above queries running. You can see that a huge amount of the profile is dominated by the tokio runtime and various atomic operations over these channels.

It is hard to see where the query execution in either the execution engines or the query engine is happening :-)

<img width="1653" alt="Screenshot 2021-04-28 at 12 09 28" src="https://user-images.githubusercontent.com/501993/116394313-aa9d8c80-a81a-11eb-9b2d-1ea4aa0a1fed.png">

## Approach

Effectively this PR removes these intermediate channels and opts to buffer up materialised results before they're sent out. There are pros and cons to this approach:

 * It's significantly more performant than the current implementation;
 * It's more performant that when I manually compiled in "tuned" buffer sizes for all the channels;
 * The code is conceptually simpler with less concurrency;
 * However, I believe that this change will effectively mean there are fewer yield points and so the query execution may hog a thread for longer;
 * Also, memory usage may increase because we will likely be buffering more data before we send it out to clients.

Overall I feel that it's preferable to start here with conceptually simpler code and then in the future if we determine we are hitting problems caused by the downsides above we can add some yielding/back-pressure in. I think it will be easier to do it that way around now that we have some good workloads running through IOx.

Finally, async Rust code is not something I am very familiar with yet so I am more than happy for others to contribute improvements either as extra commits on this PR or by opening up following PRs.

## Benchmark Results

Here are some full benchmark results:

On my intel `i9` machine:

```
                                                                                                                   THIS PR                   MAIN
read_filter/Data in one open chunk and one closed chunk of mutable buffer/no_pred                         1.00     53.7±5.18ms        4.60   247.1±32.04ms   
read_filter/Data in one open chunk and one closed chunk of mutable buffer/with_pred_tag_3=value49         1.00  1789.8±134.48µs       2.89      5.2±0.11ms   
read_filter/Data in open chunk of mutable buffer, and one chunk of read buffer/no_pred                    1.00     52.7±5.50ms        4.41   232.3±22.90ms   
read_filter/Data in open chunk of mutable buffer, and one chunk of read buffer/with_pred_tag_3=value49    1.00      3.3±0.16ms        2.06      6.7±0.21ms   
read_filter/Data in single open chunk of mutable buffer/no_pred                                           1.00     28.5±2.25ms        4.00    114.0±3.65ms   
read_filter/Data in single open chunk of mutable buffer/with_pred_tag_3=value49                           1.00  1328.3±133.32µs       2.27      3.0±0.10ms   
read_filter/Data in two read buffer chunks/no_pred                                                        1.00     49.7±4.54ms        4.56   226.8±11.68ms   
read_filter/Data in two read buffer chunks/with_pred_tag_3=value49                                        1.00      4.7±0.18ms        1.79      8.4±0.18ms   
read_group/Data in one open chunk and one closed chunk of mutable buffer/no_pred                          1.00   439.0±16.22ms        1.32   578.7±87.27ms   
read_group/Data in one open chunk and one closed chunk of mutable buffer/with_pred_tag_3=value49          1.00      7.6±0.46ms        1.34     10.1±0.38ms   
read_group/Data in open chunk of mutable buffer, and one chunk of read buffer/no_pred                     1.00   446.6±19.03ms        1.25   559.1±74.64ms   
read_group/Data in open chunk of mutable buffer, and one chunk of read buffer/with_pred_tag_3=value49     1.00      9.0±0.50ms        1.29     11.6±0.49ms   
read_group/Data in single open chunk of mutable buffer/no_pred                                            1.00   436.6±16.47ms        1.18   516.1±22.78ms   
read_group/Data in single open chunk of mutable buffer/with_pred_tag_3=value49                            1.00      7.3±0.28ms        1.38     10.1±0.42ms   
read_group/Data in two read buffer chunks/no_pred                                                         1.00   450.7±18.58ms        1.18   533.5±17.60ms   
read_group/Data in two read buffer chunks/with_pred_tag_3=value49                                         1.00     10.7±0.65ms        1.24     13.3±0.35ms   
```

This PR's benchmarks are on the left column and `main` is on the **last** column. These results show that performance on these benchmarks has improved between `18–366%` depending on the exact benchmark.


And also on `aarch64` using an Mac Mini with M1 chip.

```
                                                                                                                   THIS PR                   MAIN
read_filter/Data in one open chunk and one closed chunk of mutable buffer/no_pred                         1.00     27.9±0.53ms       28.00  782.3±66.85ms
read_filter/Data in one open chunk and one closed chunk of mutable buffer/with_pred_tag_3=value49         1.00  1550.0±208.27µs      10.93    16.9±2.30ms
read_filter/Data in open chunk of mutable buffer, and one chunk of read buffer/no_pred                    1.00     26.7±0.63ms       29.81  795.1±57.12ms
read_filter/Data in open chunk of mutable buffer, and one chunk of read buffer/with_pred_tag_3=value49    1.00      3.1±0.13ms       6.12     18.8±1.93ms
read_filter/Data in single open chunk of mutable buffer/no_pred                                           1.00     16.1±0.43ms       24.70  397.8±45.22ms
read_filter/Data in single open chunk of mutable buffer/with_pred_tag_3=value49                           1.00  1226.1±206.57µs      7.54      9.2±1.02ms
read_filter/Data in two read buffer chunks/no_pred                                                        1.00     25.3±0.48ms       31.60  798.9±50.18ms
read_filter/Data in two read buffer chunks/with_pred_tag_3=value49                                        1.00      4.5±0.18ms       4.03     17.9±2.29ms
read_group/Data in one open chunk and one closed chunk of mutable buffer/no_pred                          1.00    188.8±8.36ms       3.19   601.4±20.85ms
read_group/Data in one open chunk and one closed chunk of mutable buffer/with_pred_tag_3=value49          1.00      4.7±0.19ms       2.85     13.3±1.60ms
read_group/Data in open chunk of mutable buffer, and one chunk of read buffer/no_pred                     1.00    188.0±8.01ms       3.17   595.9±31.51ms
read_group/Data in open chunk of mutable buffer, and one chunk of read buffer/with_pred_tag_3=value49     1.00      5.9±0.15ms       2.42     14.3±1.17ms
read_group/Data in single open chunk of mutable buffer/no_pred                                            1.00    182.6±8.58ms       3.25   592.5±22.80ms
read_group/Data in single open chunk of mutable buffer/with_pred_tag_3=value49                            1.00      4.6±0.19ms       2.89     13.2±0.89ms
read_group/Data in two read buffer chunks/no_pred                                                         1.00    191.6±9.67ms       3.02   578.9±45.98ms
read_group/Data in two read buffer chunks/with_pred_tag_3=value49                                         1.00      7.2±0.18ms       2.23     16.1±1.15ms
```

The interesting thing about the `aarch64` results is there is a _very large_ improvement in benchmark times, e.g. `~3000%` in one case. This is possiblly due to intel having a better architecture for dealing with the type of contention/synchronisation that was having to happen with the channel-based implementation. 

With this PR you can see that the benchamarks improve to the point where they're faster on `aarch64` than `x86`.

We should consider that the changes in this PR made need to be eventually brought back more towards the existing approach if we need to add more yielding and less buffering, but I think it's better to start from here and go in that direction.